+++ packages/api/src/client/types/app/bsky/unspecced/getSuggestionsSkeleton.ts
@@ -6,18 +6,20 @@ import { ValidationResult, BlobRef } from '@atproto/lexicon'
 import { isObj, hasProp } from '../../../../util'
 import { lexicons } from '../../../../lexicons'
 import { CID } from 'multiformats/cid'
+import * as AppBskyUnspeccedDefs from './defs'
 
-export interface QueryParams {}
+export interface QueryParams {
+  /** DID of the account making the request (not included for public/unauthenticated queries). Used to boost followed accounts in ranking. */
+  viewer?: string
+  limit?: number
+  cursor?: string
+}
 
 export type InputSchema = undefined
 
 export interface OutputSchema {
-  handle: string
-  did: string
-  email?: string
-  emailConfirmed?: boolean
-  emailAuthFactor?: boolean
-  didDoc?: {}
+  cursor?: string
+  actors: AppBskyUnspeccedDefs.SkeletonSearchActor[]
   [k: string]: unknown
 }